+++ src/CoordinateReferenceSystem/Geocentric.php
@@ -1486,9 +1486,17 @@ class Geocentric extends CoordinateReferenceSystem
     /**
      * WGS 84 (G1762)
      * Extent: World.
-     * Replaces WGS 84 (G1674) (CRS code 7662) from 2013-10-16.
+     * Replaces WGS 84 (G1674) (CRS code 7662) from 2013-10-16. Redesignated WGS 84 (G1762') in 2015 after changes to 7
+     * NGA tracking station locations and antennas. Replaced by WGS 84 (G2139) (CRS code 9753) from 2021-01-03.
      */
     public const EPSG_WGS_84_G1762 = 'urn:ogc:def:crs:EPSG::7664';
+
+    /**
+     * WGS 84 (G2139)
+     * Extent: World.
+     * Replaces WGS 84 (G1762) (CRS code 7664) from 2021-01-03.
+     */
+    public const EPSG_WGS_84_G2139 = 'urn:ogc:def:crs:EPSG::9753';
 
     /**
      * WGS 84 (G730)