+++ src/CoordinateReferenceSystem/Geographic3D.php
@@ -1557,9 +1557,17 @@ class Geographic3D extends Geographic
     /**
      * WGS 84 (G1762)
      * Extent: World.
-     * Replaces WGS 84 (G1674) (CRS code 7663) from 2013-10-16.
+     * Replaces WGS 84 (G1674) (CRS code 7663) from 2013-10-16. Redesignated WGS 84 (G1762') in 2015 after changes to 7
+     * NGA tracking station locations and antennas. Replaced by WGS 84 (G2139) (CRS code 9754) from 2021-01-03.
      */
     public const EPSG_WGS_84_G1762 = 'urn:ogc:def:crs:EPSG::7665';
+
+    /**
+     * WGS 84 (G2139)
+     * Extent: World.
+     * Replaces WGS 84 (G1762) (CRS code 7665) from 2021-01-03.
+     */
+    public const EPSG_WGS_84_G2139 = 'urn:ogc:def:crs:EPSG::9754';
 
     /**
      * WGS 84 (G730)